+++ src/Instagram.php
@@ -1466,9 +1466,9 @@ class Instagram implements ExperimentsInterface
             if ($this->loginAttemptCount === 0 && !self::$skipLoginFlowAtMyOwnRisk) {
                 $this->_sendPreLoginFlow();
             } else {
-                $launcherResponse = $this->internal->sendLauncherSync(true)->getHttpResponse();
-                $this->settings->set('public_key', $launcherResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
-                $this->settings->set('public_key_id', $launcherResponse->getHeaderLine('ig-set-password-encryption-key-id'));
+                $mobileConfigResponse = $this->internal->getMobileConfig(true)->getHttpResponse();
+                $this->settings->set('public_key', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
+                $this->settings->set('public_key_id', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-key-id'));
             }
 
             $this->event->sendStringImpressions(['2131231876' => 1, '2131231882' => 1, '2131886885' => 2, '2131887195' => 1, '2131887196' => 1, '2131888193' => 4, '2131888472' => 1, '2131890367' => 1, '2131891325' => 1, '2131892179' => 1, '2131892669' => 1, '2131892673' => 1, '2131893765' => 1, '2131893766' => 1, '2131893767' => 1, '2131893768' => 1, '2131893769' => 1, '2131893770' => 1, '2131893771' => 1, '2131893772' => 1, '2131893773' => 1, '2131893774' => 1, '2131893775' => 1, '2131893776' => 1, '2131893777' => 1, '2131893778' => 1, '2131893779' => 1, '2131893780' => 1, '2131893781' => 1, '2131893782' => 1, '2131893783' => 1, '2131893784' => 1, '2131893785' => 1, '2131893788' => 1, '2131893789' => 1, '2131893790' => 1, '2131893791' => 2, '2131893792' => 1, '2131893793' => 1, '2131893806' => 1, '2131893898' => 1, '2131894010' => 1, '2131894018' => 1, '2131896911' => 1, '2131898165' => 1]);
@@ -2293,9 +2293,9 @@ class Instagram implements ExperimentsInterface
                     // pass
                 }
                 */
-                $launcherResponse = $this->internal->sendLauncherSync(true)->getHttpResponse();
-                $this->settings->set('public_key', $launcherResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
-                $this->settings->set('public_key_id', $launcherResponse->getHeaderLine('ig-set-password-encryption-key-id'));
+                $mobileConfigResponse = $this->internal->getMobileConfig(true)->getHttpResponse();
+                $this->settings->set('public_key', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
+                $this->settings->set('public_key_id', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-key-id'));
                 //$this->internal->bootstrapMsisdnHeader();
                 try {
                     $this->internal->logAttribution();
@@ -2311,9 +2311,9 @@ class Instagram implements ExperimentsInterface
             // Start emulating batch requests with Pidgeon Raw Client Time.
             $this->client->startEmulatingBatch();
         } else {
-            $launcherResponse = $this->internal->sendLauncherSync(true)->getHttpResponse();
-            $this->settings->set('public_key', $launcherResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
-            $this->settings->set('public_key_id', $launcherResponse->getHeaderLine('ig-set-password-encryption-key-id'));
+            $mobileConfigResponse = $this->internal->getMobileConfig(true)->getHttpResponse();
+            $this->settings->set('public_key', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
+            $this->settings->set('public_key_id', $mobileConfigResponse->getHeaderLine('ig-set-password-encryption-key-id'));
         }
 
         try {
@@ -2329,7 +2329,7 @@ class Instagram implements ExperimentsInterface
             if ($this->getPlatform() === 'ios') {
                 $this->account->getNamePrefill();
             }
-            $this->internal->sendLauncherSync(true, true, true);
+            $this->internal->getMobileConfig(true);
 
             /* QE SYNC DISABLED
             try {
@@ -2453,17 +2453,9 @@ class Instagram implements ExperimentsInterface
 
             try {
                 $this->account->getAccountFamily();
-                $this->internal->sendLauncherSync(false, false, true);
+                $this->internal->getMobileConfig(false);
                 $this->internal->fetchZeroRatingToken();
                 $this->event->sendZeroCarrierSignal();
-
-                /* DISABLED. IG DOES NOT SEND THESE ANYMORE.
-                try {
-                    $this->internal->syncUserFeatures();
-                } catch (\Exception $e) {
-                    // pass
-                }
-                */
             } finally {
                 // Stops emulating batch requests.
                 $this->client->stopEmulatingBatch();
@@ -2532,7 +2524,7 @@ class Instagram implements ExperimentsInterface
             $this->client->startEmulatingBatch();
 
             try {
-                $this->internal->sendLauncherSync(false, false, true, true);
+                $this->internal->getMobileConfig(false);
                 $this->story->getReelsMediaFeed($this->account_id);
             } finally {
                 // Stops emulating batch requests
@@ -2769,20 +2761,6 @@ class Instagram implements ExperimentsInterface
             if ($lastExperimentsTime === null || (time() - $lastExperimentsTime) > self::EXPERIMENTS_REFRESH) {
                 // Start emulating batch requests with Pidgeon Raw Client Time.
                 $this->client->startEmulatingBatch();
-
-                /* QE SYNC DISABLED
-                try {
-                    $this->internal->syncUserFeatures();
-                    $this->internal->syncDeviceFeatures();
-                }
-                catch(\InstagramAPI\Exception\EmptyResponseException | \InstagramAPI\Exception\ThrottledException $e){
-                    // This can have EmptyResponse, and that's ok.
-                }
-                finally {
-                    // Stops emulating batch requests.
-                    $this->client->stopEmulatingBatch();
-                }
-                */
             }
 
             // Update zero rating token when it has been expired.